% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_tools.R
\name{addalpha}
\alias{addalpha}
\title{Add alpha to channel 4 in rgba space}
\usage{
addalpha(colors, alpha = 1)
}
\arguments{
\item{colors}{Input colors which will be converted to rgb}

\item{alpha}{Set constant alpha level for rgb colors}
}
\value{
rbga colors
}
\description{
Add alpha to channel 4 in rgba space
}
