% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_pattern_alignment.R
\name{align.grids}
\alias{align.grids}
\title{Align grid using shiny app}
\usage{
align.grids(scatter, df_ref = NULL, plot.coords = F, return.df = T,
  alpha = 0.5, size = 0.5, output.name = "trans.coord.list",
  keep.ref.fixed = F, scale.coords.ind = F)
}
\arguments{
\item{scatter}{A dot pattern with x, y values or a list with dot pattern in as the first element and corresponding coordinates
as the second element.}

\item{df_ref}{list with dot pattern in as the first element and corresponding coordinates
as the second element used for alignment. The reference pattern can be obtained by running this function
without a reference dot pattern.}

\item{plot.coords}{Logical specifying if feature coordinates should be plotted on top of binary dot pattern.
For the first replicate, it is beneficial to set this to true to evaluate how well the features align with the dot pattern.}

\item{return.df}{Logical specifying if the manipulated data should be returned.}

\item{alpha}{Numeric value specifying alpha level of feature spots. Only required if plot.coords = TRUE.}

\item{size}{Numeric value specifying size of feature spots. Only required if plot.coords = TRUE.}

\item{output.name}{Character specifying the name of the output object.
by running the function without df_ref.}
}
\value{
Download normalized data to an .RData file.
}
\description{
Align binary dot pattern and calculate normalized coordinates. The alignment is doe using a local shiny app.
A reference dot pattern (black color) serves as a template for the alignment of a second dot pattern (red color). The second
dot pattern is automatically scaled to center around 0 and can then be rotated, scaled along x and y axis, flipped or moved.
Once the dot pattern is properly aligned with the reference dot pattern, the data can be downloaded as an .RData file.
}
