% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\docType{methods}
\name{batch.correct}
\alias{batch.correct}
\alias{batch.correct,spaceST-method}
\title{Batch correction of spaceST data.}
\usage{
batch.correct(object)

\S4method{batch.correct}{spaceST}(object)
}
\arguments{
\item{object}{Object of class spaceST.}
}
\value{
Batch corrected data frame in slot corrected
}
\description{
Run batch correction of replicate sections.
}
\details{
Function used to run a batch correction between samples.
}
\seealso{
\link[countClust]{BatchCorrectedCounts}
}
