% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_bootstrap.R
\docType{methods}
\name{bias.bootstraps,xCellBoot-method}
\alias{bias.bootstraps,xCellBoot-method}
\title{Extract bias data}
\usage{
\S4method{bias.bootstraps}{xCellBoot}(object)
}
\value{
Data.frame with celltypes as columns and replicates as rows.
}
\description{
Extract the median xCell scores for each celltype and cluster from an xCellBoot object.
}
