% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cast2list}
\alias{cast2list}
\title{Cast merged data to list of matrixes}
\usage{
cast2list(df)
}
\arguments{
\item{df}{Merged data.frame/matrix with spaceST headers, i.e. features as columns.}
}
\value{
List of matrices.
}
\description{
Cast any merged data.frame/matrix into a list of matrices for each replicate.
}
