% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_bootstrap.R
\docType{methods}
\name{celltype.bootstraps,xCellBoot-method}
\alias{celltype.bootstraps,xCellBoot-method}
\title{Extract celltype bootstrap replicates}
\usage{
\S4method{celltype.bootstraps}{xCellBoot}(object, celltype)
}
\arguments{
\item{celltype}{Character string specifying xCell signature celltype.}
}
\value{
Data.frame with clusters as columns and replicates as rows.
}
\description{
Extract the bootstrap results for a specific celltype.
}
