% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_tools.R
\name{clust2feature}
\alias{clust2feature}
\title{Rearrange xCell output from clusters to features}
\usage{
clust2feature(xCell.matrix, clusters, features)
}
\arguments{
\item{clusters}{Integer vector specifying cluster positions.}

\item{features}{Character vector with feature names of the original data.frame that was used for pooling.}
}
\value{
A rearranged xCell matrix with features as columns and genes as rows.
}
\description{
This function can rearrange the clusters in the xCell score matrix back into features.
During the rearrangement, the cluster score is assigned to each feature within that cluster. This assumes
that every feature within each cluster has the same celltype mixture, which is a rough approximation.
However, high xCell scores indicate a strong presence of a certain celltype in that cluster even though
the exact celltype positioning is unknown.
}
