% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_bootstrap.R
\docType{methods}
\name{cluster.bootstraps,xCellBoot-method}
\alias{cluster.bootstraps,xCellBoot-method}
\title{Extract cluster bootstrap replicates}
\usage{
\S4method{cluster.bootstraps}{xCellBoot}(object, cluster, filter = F,
  probs = 0.1, threshold = 0.001)
}
\arguments{
\item{cluster}{Integer specifying cluster number.
#' @param filter Logical specifying wether or not filtering should be applied.}

\item{probs}{Chose quantile to use for filtering.}

\item{threshold}{Numeric value specifying threshold. If filter = TRUE, columns with probs_quantile > threshold will be kept.}
}
\value{
Data.frame with celltypes as columns and replicates as rows.
}
\description{
Extract the bootstrap results for a specific cluster.
}
