% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{cluster_matrix}
\alias{cluster_matrix}
\title{Compute cluster matrix.}
\usage{
cluster_matrix(object, clusters)
}
\arguments{
\item{object}{Expression data.frame, matrix or object of class spaceST.}

\item{clusters}{Integer vector specifying cluster identity of each feature.}
}
\value{
Integer vector specifying cluster identity of each feature.
}
\description{
This function is used to pool clustered features by adding the gene expression values within each cluster.
}
