% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{clusterST}
\alias{clusterST}
\title{Calculate clusters based on topics.}
\usage{
clusterST(object = NULL, method.dist = "euclidean",
  method.tree = "ward.D2", minClusterSize = 30, custom.matrix = NULL)
}
\arguments{
\item{object}{Object of class spaceST.}

\item{method.dist}{Set distance method.}

\item{method.tree}{Set clustering method.}

\item{minClusterSize}{Integer value specifying the minimum cluster size allowed.}

\item{custom.matrix}{Provide a matrix to perform hierarchical clustering on.}
}
\value{
Integer vector specifying cluster identity of each feature.
}
\description{
This function is used to cluster features based on a topics matrix.
}
