% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{collect_gene_data}
\alias{collect_gene_data}
\title{Collect data from bioMmRt}
\usage{
collect_gene_data(gene.list, organism = "hsapiens",
  filters = "ensembl_gene_id_version",
  attributes = c("ensembl_gene_id_version", "gene_biotype"))
}
\arguments{
\item{gene.list}{character vector of gene ids}

\item{organism}{select organism (default hsapiens)}

\item{filters}{select filters from input data.frame}

\item{attributes}{select attributes to collect from biomaRt}
}
\description{
function used to collect data from biomaRt and save to data.frame
}
