% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{convert_coordinates_from_json}
\alias{convert_coordinates_from_json}
\title{Convert coordinates produced from json}
\usage{
convert_coordinates_from_json(df, rep)
}
\arguments{
\item{df}{Dataframe with json headers.}

\item{rep}{Integer value specifying replicate number.}
}
\value{
Return data.frame with modified headers.
}
\description{
Use this funciton to convert coordinate names starting with X and add replicate
number.
}
