% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_coordinates}
\alias{get_coordinates}
\alias{get_coordinates.matrix}
\alias{get_coordinates.data.frame}
\alias{get_coordinates.spaceST}
\alias{get_coordinates.list}
\title{get_coordinates: obtain spatial coordinates}
\usage{
get_coordinates(x, rep = TRUE)

\method{get_coordinates}{matrix}(x, rep = TRUE)

\method{get_coordinates}{data.frame}(x, rep = TRUE)

\method{get_coordinates}{spaceST}(x, rep = TRUE)

\method{get_coordinates}{list}(x, rep = TRUE)
}
\arguments{
\item{x}{List of expression data.frames, data.frame or matrix.}

\item{rep}{Get replicate column}
}
\value{
Data frame with replicate numbers, x and y coordinates
}
\description{
This function is used to collect all coordinates from a list of expression datasets or an expression dataset
}
\details{
Obtain coordinates from expression data list of expression data.frame
}
