% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_bootstrap.R
\name{CreatexCellBoot}
\alias{CreatexCellBoot}
\title{Bootstrap xCell data.}
\usage{
CreatexCellBoot(data, clusters, replicates)
}
\arguments{
\item{data}{Input data.frame.}

\item{clusters}{Integer vector specifying cluster positions in data.frame.}

\item{replicates}{Integer specifying number of bootstrap replicates.}
}
\value{
An xCellBoot object.
}
\description{
Used to bootstrap xCell analysis results.
}
