% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{DE_heatmap}
\alias{DE_heatmap}
\title{Plot DE heatmap}
\usage{
DE_heatmap(object, FDR.cutoff = 0.05, logFC.cutoff = 2, log.2 = T,
  type = "normalized", include.downregulated = F, scale.data = T)
}
\arguments{
\item{object}{Object of class spaceST with DE results.}

\item{FDR.cutoff}{Cutoff for adjusted p-values.}

\item{logFC.cutoff}{Cutoff for logFC.}

\item{log.2}{Log2 transform data.}

\item{type}{Select data type from spaceST object.}

\item{include.downregulated}{Include downregulated genes when selecting gene to visualize in heatmap.
Absolute value will be used for the logFC threshold.}

\item{scale.data}{Logical speficying whether or not values should be scaled.}
}
\value{
Heatmap of DE genes.
}
\description{
Function used to plot heatmaps of DE genes expression
}
