% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ensembl2hgnc}
\alias{ensembl2hgnc}
\alias{ensembl2hgnc.data.frame}
\alias{ensembl2hgnc.matrix}
\alias{ensembl2hgnc.list}
\title{Convert from ENSEMBL ID to HGNC symbol}
\usage{
ensembl2hgnc(df)

\method{ensembl2hgnc}{data.frame}(df)

\method{ensembl2hgnc}{matrix}(df)

\method{ensembl2hgnc}{list}(df)
}
\arguments{
\item{df}{Data.frame or matrix with ENSEMBL ids as rownames.}
}
\value{
Data.frame or matrix with converted names.
}
\description{
Function used to convert ENSEMBL gene ids of an expression matrix into HGNC symbols.
}
