% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{ExtractTopFeatures}
\alias{ExtractTopFeatures}
\title{Extract top features}
\usage{
ExtractTopFeatures(object, top_features = 1000, method = "poisson",
  options = "min", shared = FALSE)
}
\arguments{
\item{object}{Object of class spaceST.}
}
\description{
Extract top features
}
