% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{ExtractTopFeaturesST}
\alias{ExtractTopFeaturesST}
\title{Extract top features}
\usage{
ExtractTopFeaturesST(object, top_features = 1000, method = "poisson",
  options = "min", shared = FALSE)
}
\arguments{
\item{object}{Object of class spaceST.}
}
\description{
Extract top features
}
\seealso{
\link[CountClust]{ExtractTopFeatures}
}
