% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\name{filter}
\alias{filter}
\title{Method filter for spaceST class.}
\usage{
filter(object, unique.genes = 300, min.exp = 2, min.features = 15,
  filter.genes = NULL, force.filter = FALSE, delimiter = "_")
}
\arguments{
\item{object}{Object of class spaceST.}

\item{unique.genes}{The lowest number of unique genes allowed in a feature (spot).}

\item{min.exp}{Integer value specifying the lowest expression value allowed at least min.features number of features.}

\item{min.features}{Integer value specifying the lowest number of features with at least min.exp.}

\item{filter.genes}{A character vector specifying genes that should be filtered from the expression data.}

\item{force.filter}{Force filter on a spaceST object.}

\item{delimiter}{Character used to delimit coordinates in headers.}
}
\description{
Apply filter to spaceST data.
}
