% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_tools.R
\name{format_3D_data}
\alias{format_3D_data}
\title{Used to format 3D data}
\usage{
format_3D_data(df, trans, object)
}
\arguments{
\item{df}{Data.frame with object values, e.g. xCell scores, gene expression, topics, ...}

\item{trans}{Transformed x, y coordinates centered around (0, 0).}

\item{object}{Specific object to analyze, e.g. celltype, gene, or topic.}
}
\value{
Object specific values and coordinates.
}
\description{
Function used to obtain specific object, e.g. celltype or gene, and bind values with
transformed coordinates.
}
