% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_2D_data}
\alias{interpolate_2D_data}
\title{Interpolate object values over "cell population"}
\usage{
interpolate_2D_data(val.df, section.input, x, y, iterval = NULL)
}
\arguments{
\item{val.df}{Data.frame with object values.}

\item{section.input}{Data.frame with dot pattern coordinates.}

\item{x}{Number of grids along x axis used for raster.}

\item{y}{Number of grids along y axis used for raster.}
}
\value{
data.frame with input values for feature coordinates interpolated over 2D scatter.
}
\description{
Function used to interpolate values, e.g. gene expression values, xCell scores or topic proportions
over dot pattern.
}
