% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{LDA}
\alias{LDA}
\alias{topic_compute}
\alias{topic_compute.default}
\alias{topic_compute.spaceST}
\title{Helper function to compute lda using cellTree package.}
\usage{
topic_compute(object, min.topic = 2, max.topic = 15,
  num.genes = NULL, datatype = "norm.data", method = "maptpx",
  sd.filter = F, log.scale = F, force.recalc = F,
  minClusterSize = 30)

\method{topic_compute}{default}(object, min.topic = 2, max.topic = 15,
  num.genes = NULL, datatype = "norm.data", method = "maptpx",
  sd.filter = F, log.scale = F, force.recalc = F,
  minClusterSize = 30)

\method{topic_compute}{spaceST}(object, min.topic = 2, max.topic = 15,
  num.genes = NULL, datatype = "norm.data", method = "maptpx",
  sd.filter = FALSE, log.scale = FALSE, force.recalc = FALSE,
  minClusterSize = 30)
}
\arguments{
\item{object}{Object of class spaceST.}

\item{min.topic, max.topic}{Integer values specifying min and max number of topics in the "maptpx" model.}

\item{num.genes}{Number of genes to use for lda modelling.}

\item{datatype}{Character string specifying if the "expr" or "norm.data" ata should be used as input. Default
is set to "norm.data".}

\item{method}{Method passed to CellTree function.}

\item{sd.filter}{Standard deviation threshold.}

\item{log.scale}{Convert to log scale.}
}
\value{
Topic matrix
}
\description{
This helper function is used to compute topics for an expression dataset using lda modelling from cellTree package.
}
\references{
\url{https://bioconductor.org/packages/release/bioc/html/cellTree.html}
}
\seealso{
\link[cellTree]{compute.lda}
}
