% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_pattern_alignment.R
\name{manipulate.grid}
\alias{manipulate.grid}
\title{Scale, move and rotate binary dot pattern}
\usage{
manipulate.grid(data, df_ref = NULL, sx = 1, sy = 1, shift_x = 0,
  shift_y = 0, flip_x = F, flip_y = F, alpha = 0.5, size = 0.5,
  plot.coords = F, return.df = F, keep.ref.fixed = F)
}
\arguments{
\item{data}{A list with a binary dot pattern and corresponding feature coordinates used to create the pattern.}

\item{df_ref}{A reference binary dot pattern that that you want to align the new pattern to.}

\item{sx}{Scale x coordinates with sx scalar, i.e. expand or contract along x axis.}

\item{sy}{Scale y coordinates with sy scalar, i.e. expand or contract along y axis.}

\item{shift_x}{Move pattern along x axis.}

\item{shift_y}{Move pattern along y axis.}

\item{flip_x}{Mirror along x axis.}

\item{alpha}{Set alpha level for data in plot.}

\item{size}{Set size of dots in plot.}

\item{keep.ref.fixed}{Fix the reference dot pattern if you only want to change the coordinates.}

\item{fip_y}{Mirror along y axis.}
}
\value{
Dataframe or list with scaled data.
}
\description{
This function is used to manually align binary dot patterns.
}
