% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_bootstrap.R
\docType{methods}
\name{mean.bootstraps,xCellBoot-method}
\alias{mean.bootstraps,xCellBoot-method}
\title{Extract mean bootstrap values}
\usage{
\S4method{mean.bootstraps}{xCellBoot}(object)
}
\value{
Data.frame with celltypes as columns and replicates as rows.
}
\description{
Extract the mean xCell scores for each celltype and cluster from an xCellBoot object.
}
