% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{merge_exp_list}
\alias{merge_exp_list}
\title{Merge and filter expression datasets}
\usage{
merge_exp_list(x, unique.genes = 0, min.exp = 0, min.features = 0,
  filter.genes = NULL)
}
\arguments{
\item{x}{List of exrpression data frames, data.frame or matrix.}

\item{unique.genes}{Integer value specifying number of unique genes allowed in a feature.}

\item{min.exp}{Integer value specifying lowest expression value allowed in min.features number of features.}

\item{min.features}{Integer value specifying number of features allowed with min.exp count.}

\item{filter.genes}{Character vector specifying genes that should be filtered out.}
}
\value{
Merged and filtered dataframe.
}
\description{
This function is used to merge datasets in a list of expression datasets.
Genes which are not present in some datasets will get 0 values in those datasets. The merged dataset is
subsequently filtered from features with few unique genes and genes with low expression counts.
}
