% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_tools.R
\name{morph.3d}
\alias{morph.3d}
\title{Plot interactive 3D scatter}
\usage{
morph.3d(object, exp.list, trans.list, section.input.list, max,
  colorscale = "Jet", background.col = "rgb(0, 0, 0)",
  grid.color = "rgb(0, 0, 0)", marker.size = 1.5, min.size = 0,
  max.size = 1.5, range = NULL, opacity = 0.1, show.top = NULL,
  scalefactor = 0, showtickx = FALSE, showticky = FALSE,
  showtickz = FALSE, add_sections = FALSE, size.scale.factor = 0,
  cam = NULL, skip.every = 3, return.df = FALSE, subsample = NULL,
  add.rotation = TRUE, rot.angle = 0.01, rot.offset = 20)
}
\arguments{
\item{object}{Gene name, celltype, topic etc. specifying the object you want to visualize. This will depend on the
content of exp.list which holds the values you want to visualize. "object" must be located in the row names of the
exp.list data.frames.}

\item{exp.list}{List of data.frames containing values of interest, e.g. gene expression values, topic proportions,
clusters, xCell scores etc.}

\item{trans.list}{List with transformed x, y coordinates for each replicate section.}

\item{section.input.list}{List of section.input objects, i.e. data.frames with dot pattern x, y coordinates and
corresponding grid cells.}

\item{max}{Integer value specifying maximum value of colorscale.}

\item{colorscale}{Set colorscale for 3D scatter (e.g. Jet, Portland, Viridis and grey, see plotly documentation)}

\item{background.col}{Sets background color using rgb channels. Default 'rgb(0, 0, 0)' = "black". Use function col2rgb("color") to
obtain rgb channel values for custom colors ("white" = 'rgb(255, 255, 255)).}

\item{grid.color}{Set color of background grid.}

\item{marker.size}{Numeric value specifying marker size (point size).}

\item{min.size}{Set minimum size of markers.}

\item{max.size}{Set maximum size of markers.}

\item{range}{Set range of colorscale. range = "zero_to_one" will scale values to [0, 1] interval.}

\item{opacity}{Numeric between 0-1 specifying marker opacity.}

\item{show.top}{Numeric between 0-1 specifying which subset of data to show. show.top = 0.5 will show all markers with
values larger than 50*max(value).}

\item{scalefactor}{Integer which adds empty distance outside tissue section along z axis. Larger value will make sections
appear closer to each other.}

\item{showtickx}{Logical specifying whther or not to show x axis values.}

\item{showticky}{Logical specifying whther or not to show y axis values.}

\item{showtickz}{Logical specifying whther or not to show z axis values.}

\item{add_sections}{Repeat tissue sections 3 times to increase 3D effect. NOTE: 60% of the data
points in each section will be removed to improve performance}

\item{size.scale.factor}{use value parameter (e.g. xCell score) to set marker size for each dot. size.scale.factor
is used as a multiplier, i.e. {markersize = value*scalefactor} for each datapoint. For example, xCell scores
ranging between 0 -> 0.2 can be scaled up to a maximum marker size of 1 by using a scalefactor of 5. Scaling
the dots this way will make datapoints with low values less visible and make the high value volumes more distinct
in the scatter plot. This scaling can be used in combination with show.top to clean "less important" space from
data points which are concealing the high value clusters.}

\item{cam}{Set camera angle of plot object.}

\item{skip.every}{Used to subset final data output.}

\item{return.df}{Logical specifying whether or not the data should be returned. This will override the plot functionality.}

\item{subsample}{Subsample data to decrease number of markers.}

\item{add.rotation}{Add rotation button to html widget.}

\item{rot.angle}{Set roation angle of camera.}

\item{rot.offset}{Set rotation offset of camera.}
}
\value{
Interactive 3D scatter object.
}
\description{
Function used to create a 3D scatter of ST data using consecutive tissue sections.
}
