% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_cpm}
\alias{calc_cpm}
\title{Normalize using CPTK method}
\usage{
calc_cpm(expr_mat)
}
\arguments{
\item{expr_mat}{Expression data.frame or matrix.}
}
\value{
Normalized data.frame or matrix.
}
\description{
Function used to normalize ST data using Counts Per Ten thousand method
}
