% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST.R
\name{NormalizespaceST}
\alias{NormalizespaceST}
\title{Normalize ST data}
\usage{
NormalizespaceST(object, method = "scran", log2 = F, clusters = NULL, ...)
}
\arguments{
\item{object}{spaceST object with expression data to correct}

\item{method}{select normalization method, default "cp10k" (options: "cp10k", "scran")}

\item{log2}{Logical specifying whether or not data should be log2-transformed}

\item{clusters}{Integer vector specifying clusters used for normalization of large
sce objects. Only for "scran" normalization.}
}
\value{
Normalized data frame in slot normalized
}
\description{
Normalize ST data
}
