% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_bootstrap.R
\docType{methods}
\name{obs.bootstraps,xCellBoot-method}
\alias{obs.bootstraps,xCellBoot-method}
\title{Extract obserevd bootstrap values}
\usage{
\S4method{obs.bootstraps}{xCellBoot}(object, filter = F, probs = 0.1,
  threshold = 0.001)
}
\arguments{
\item{filter}{Logical specifying whether or not the data should be filtered from unreliable estimates.}

\item{probs}{Numerical value between 0-1 specifying the quantile used for filtering. (see threshold)}

\item{threshold}{Set threshold for xCell score for filtering. For a given celltype distribution of xCell scores,
the quantile value specified by the probs parameter will be used as a test to decide if the estimate should be kept or not.
The xCell score distributions might be heavily skewed with a large number of 0 values, which could arise due to the fact that
the celltype gene signature used to calculate those estimated are only present in a subset of the input features.
With the default filter settings, the 0.1 quantile is calculated for each distribution and if this value is larger
than the threshold of 0.001, the estimate remains unchanged. If the value is lower than the threshold, the
estimate is set to 0 instead. Distributions failing the test usually contain very low xCell scores but if the estimate
is too biased, this filtering serves as a means of omitting unreliable results.}
}
\value{
Observed xCell scores for clusters.
}
\description{
Extract the observed xCell scores for each celltype and cluster from an xCellBoot object.
}
