% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST.R
\docType{methods}
\name{pca.spaceST}
\alias{pca.spaceST}
\alias{pca.spaceST,spaceST-method}
\title{Plot pca of spaceST data}
\usage{
pca.spaceST(object, ...)

\S4method{pca.spaceST}{spaceST}(object, ...)
}
\arguments{
\item{object}{spaceST object with expression data}

\item{...}{arguments passed to pca_plot}
}
\value{
Plot of the first two principal components of expression data.
}
\description{
This function is used to plot the first two principal components of expression data
stored in a spaceST object. By default, both raw expression and batch corrected exåression
data will be plotted if both are present.
}
