% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST.R
\name{pca.spaceST}
\alias{pca.spaceST}
\title{Plot pca of spaceST data}
\usage{
pca.spaceST(object, ...)
}
\arguments{
\item{object}{spaceST object with expression data}
}
\value{
Plot of the first two principal components of expression data.
}
\description{
This function is used to plot the first two principal components of expression data
stored in a spaceST object. By default, both raw expression and batch corrected exåression
data will be plotted if both are present.
}
\examples{
library(spaceST)
data(bcST)

# Create spaceST object from gene expression data
ST.object <- CreatespaceSTobject(bcST)
# Plot PCA
pca.spaceST(ST.object)

# Or if batch corrected, `pca.spaceST()` will plot a comparison between raw and corrected PCA.
ST.oject <- batch.correct(ST.object)
pca.spaceST(ST.object)
}
