% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pca_plot}
\alias{pca_plot}
\title{pca of expression data}
\usage{
pca_plot(df1, df2 = NULL, samples, ...)
}
\arguments{
\item{df1}{Expression dataset 1}

\item{df2}{Expression dataset 2}

\item{samples}{Vector of replicate numbers of length ncol(filtered)}

\item{...}{arguments passed to plotPCA function from scater}
}
\value{
Scatter plot of PC1 and PC2.
}
\description{
This function is used to  the first two principal components of an expression dataset, or a comparison
between raw expression and corrected expression data.
}
