% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\name{plot.QC.spaceST}
\alias{plot.QC.spaceST}
\title{Plot unique genes per feature and transcripts per feature.}
\usage{
plot.QC.spaceST(object, separate = F, bins = 20, ...)
}
\arguments{
\item{object}{Object of class spaceST.}

\item{separate}{Logical indicating whether replicates should be ploteed separately.}

\item{bins}{Select number of bins.}

\item{...}{Parameters passed to geom_histogram.}
}
\value{
Histograms of unique genes per feature and transcripts per feature distributions.
}
\description{
This function is used to plot the unique genes per feature and transcipts per feature
distributions as histograms. By default, the function will take the batch corrected dataset
if present.
}
