% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\name{plotPCA}
\alias{plotPCA}
\title{Method plotPCA.}
\usage{
plotPCA(object, components = c(1, 2), ...)
}
\arguments{
\item{object}{Object of class spaceST.}

\item{components}{Integer or character vector of length 2 specifying two PC components to compare.}
}
\value{
PCA plot.
}
\description{
Plot principal components.
}
