% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST.R
\docType{methods}
\name{plot.QC.spaceST}
\alias{plot.QC.spaceST}
\alias{plot.QC.spaceST,spaceST-method}
\title{Plot unique genes per feature and transcripts per feature}
\usage{
plot.QC.spaceST(object, separate = F)

\S4method{plot.QC.spaceST}{spaceST}(object, separate = F)
}
\arguments{
\item{object}{spaceST object with expression data.}

\item{separate}{Logical indicating whether replicates should be ploteed separ2ately}
}
\value{
Histograms of unique genes per feature and transcripts per feature distributions.
}
\description{
This funciton is used to plot the unique genes per feature and transcipts per feature
distributions as histograms. By default, the function will take the batch corrected dataset
if present.
}
