% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{rasterize_scatter}
\alias{rasterize_scatter}
\title{rasterize scatter data from HE image}
\usage{
rasterize_scatter(scatter, coords, plot = FALSE, raster_offset = 0.5)
}
\arguments{
\item{scatter}{Data.frame with binary dot pattern.}

\item{coords}{Coordinate data.frame with x, y values. Use `get_coordinates()` function to obtain feature coordinates for a
specific replicate.}
}
\value{
data.frame with x, y coordinates and grid cell numbers. Column names for x, y coordinates gives the dimensions of the grid.
}
\description{
This function is used to rasterize scatter data from a black and white HE image generated with function `scatter_HE()`.
}
