% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rename_list}
\alias{rename_list}
\title{Rename rownames of lists}
\usage{
rename_list(x, ref.list, type = "rownames", ref.list.type = "colnames")
}
\arguments{
\item{x}{input list to rename element for.}

\item{ref.list}{List to take names from.}

\item{type}{Character specifying which names of x to change.}

\item{ref.list.type}{Character specifying which names of ref.list to change.}
}
\description{
Rename colnames or rownames of list.
}
