% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{RunDEspaceST}
\alias{RunDEspaceST}
\title{Run DE analysis using the edgeR package to detect differentially expressed genes in each cluster}
\usage{
RunDEspaceST(spST, clusters = "topic", verbose = TRUE, pvalue = 0.05)
}
\arguments{
\item{spST}{spaceST object}

\item{clusters}{Integer vector of cluster IDs or character specifying existing cluster vector ("topic", "SNN")}

\item{verbose}{Print messages}

\item{pvalue}{pvalue cutoff}
}
\value{
Topic matrix
}
\description{
Run DE analysis for each cluster of spots vs remaining spots
}
