% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_tools.R
\name{scale2range}
\alias{scale2range}
\title{Scale to custom range}
\usage{
scale2range(x, a, b)
}
\arguments{
\item{x}{Input numeric vector.}

\item{a}{Lower limit.}

\item{b}{Upper limit.}
}
\value{
Scaled numeric vector.
}
\description{
Scale numeric vector to fit a certain interval.
}
\examples{
library(spaceST)

# Scale vector to fit an interval between 0 and 1
x <- seq(1, 100)
scaled.x <- scale2range(x, a = 0, b = 1)
}
