% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNN.R
\name{SetIfNull}
\alias{SetIfNull}
\title{Helper function}
\usage{
SetIfNull(x, default)
}
\arguments{
\item{x}{R object}

\item{default}{R object to replace x with if x is NULL}
}
\value{
default if x is NULL
}
\description{
Helper function
}
