% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST.R
\docType{package}
\name{spaceST-package}
\alias{spaceST}
\alias{spaceST-package}
\title{This package includes functions used to process, analyze and visualize Spatial Transcriptomics data}
\description{
More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
