% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\docType{class}
\name{spaceST}
\alias{spaceST}
\alias{spaceST-class}
\title{The spaceST class.}
\description{
An S4 class object used to filter and store spatial transcriptomics data from
multiple tissue sections
}
\section{Slots}{

\describe{
\item{\code{expr}}{Object of class dcGMatrix containing gene expression values.}

\item{\code{norm.data}}{Object of class dcGMatrix containing normalized expression values.}

\item{\code{lda.results}}{Object of class topics obtained using the \link[cellTree]{compute.lda} function from cellTree.}

\item{\code{tsne}}{Matrix object representation of t-SNE results.}

\item{\code{snn}}{Matrix object representation of the SNN graph.}

\item{\code{coordinates}}{Data.frame representation of section ID and array coordinates.}

\item{\code{filter.settings}}{List of filtering settings.}

\item{\code{meta.data}}{List of meta data.}

\item{\code{version}}{package version.}
}}

