% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,spaceST-methods}
\title{Method show for spaceST class.}
\usage{
\S4method{show}{spaceST}(object)
}
\arguments{
\item{object}{Object of class spaceST.}
}
\description{
Method show for spaceST class.
}
