% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_tools.R
\name{spatial.heatmap}
\alias{spatial.heatmap}
\title{Visualize spatial heatmaps}
\usage{
spatial.heatmap(x, coords, raster, HE = NULL, alpha = 0.5, size = 1,
  mirror_x = F, mirror_y = F, title = NULL, lim_x = c(0, 32),
  lim_y = c(0, 34), max = 1)
}
\arguments{
\item{x}{Vector with numerical values with the same length as coords. Values could represent gene expresion values,
topic proportions, xCell scores etc.}

\item{coords}{data.frame with feature x, y coordinates.}

\item{raster}{data.frame with x, y coordinates and grid cell numbers produced with the scatter_rasterize() function.}

\item{HE}{path to HE image}

\item{alpha}{set alpha level of heatmap}

\item{size}{set size of points}

\item{mirror_x}{flip along x axis}

\item{mirror_y}{flip along y axis}
}
\value{
Heatmap of spatially distributed scores.
}
\description{
Visualize spatial features on top of HE image.
}
