% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_tools.R
\name{spatial.tSNE}
\alias{spatial.tSNE}
\title{Plot t-SNE and visualize clusters}
\usage{
spatial.tSNE(df, clusters = NULL, perplexity = 15)
}
\arguments{
\item{df}{data.frame with expression values.}

\item{clusters}{Integer vector specifying cluster identity if features.}
}
\value{
A scatter plot of t-SNE results.
}
\description{
Given a gene expression data.frame, this function can be used to
run a t-SNE and visualize the results. If a cluster vector is provided, each cluster will be colored
in the output plot.
}
