% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\name{spots.under.tissue}
\alias{spots.under.tissue}
\title{Method spots.under.tissue for spaceST class.}
\usage{
spots.under.tissue(object, selection.files, delimiter = "_")
}
\arguments{
\item{object}{Object of class spaceST.}

\item{selection.files}{List of paths to selection files. The order of selections has to match the order of
sample matrices used to initiate the spaceST object.}

\item{delimiter}{Delimiter for expr headers.}
}
\description{
Subset spaceST object using spot selection tables from the ST spot detector.
}
\seealso{
\link[CreatespaceSTobject]{spaceST}
}
