% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\name{spots_under_tissue}
\alias{spots_under_tissue}
\title{Method spots_under_tissue for spaceST class.}
\usage{
spots_under_tissue(object, selection.files, delimiter = "_",
  keep.filter = T)
}
\arguments{
\item{object}{Object of class spaceST.}

\item{selection.files}{List of paths to selection files. The order of selections has to match the order of
sample matrices used to initiate the spaceST object.}

\item{delimiter}{Delimiter for expr headers.}

\item{keep.filter}{Set to FALSE if you don't want to apply the same filtering settings that were used when the
spaceST object was initiated.}
}
\description{
Subset spaceST object using spot selection tables from the ST spot detector.
}
\seealso{
\link[CreatespaceSTobject]{spaceST}
}
