% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST.R
\name{spPCA}
\alias{spPCA}
\alias{spaceST-pca}
\title{pca method for spaceST}
\usage{
spPCA(object, ntop = 500, ncomponents = 2, exprs_values = "norm.data",
  ...)
}
\arguments{
\item{object}{A spaceST object}
}
\description{
Run PC analysis of expression data
}
