% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ST_statistics.plot}
\alias{ST_statistics.plot}
\title{Quality Control plot}
\usage{
ST_statistics.plot(df, separate = F)
}
\arguments{
\item{df}{Input quality control data.frame.}

\item{separate}{Should replicates be treated separately?}
}
\value{
Plot of histograms.
}
\description{
This function plots histograms of unique genes per feature and transcripts per featrure.
}
