% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ST_statistics}
\alias{ST_statistics}
\title{Quality Control statistics}
\usage{
ST_statistics(df)
}
\arguments{
\item{df}{Input gene expression data.frame}
}
\value{
Data.frame with unique genes per feature and transcripts per feature
}
\description{
Given a data.frame with gene expression values (genes as rows and observations as columns), this function will return a
data.frame with unique genes and transcript counts for each observation, as well as the sample id
}
