% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_tools.R
\name{subset.xCell}
\alias{subset.xCell}
\title{Calculate xCell scores for cluster matrix}
\usage{
\method{subset}{xCell}(clust.df, type = "lymphoids")
}
\arguments{
\item{clust.df}{xCell data.frame with clusters or features as columns and celltypes as rows.}
}
\value{
A subset data.frame with the celltypes of interest.
}
\description{
This function is used to subset xCell results based on celltype.
}
