% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{topic_cluster_matrix}
\alias{topic_cluster_matrix}
\title{Compute cluster matrix}
\usage{
topic_cluster_matrix(df, omega = NULL, clusters = NULL)
}
\arguments{
\item{df}{Expression data used to compute topics which will be pooled using cluster vector.}

\item{omega}{Topic data.frame (omega is used to calculate clusters on the fly using the topic.clusters function)}

\item{clusters}{Integer vector specifying cluster identity of each feature (if "clusters" is specified, the "omega" parameter is not needed).}
}
\value{
Integer vector specifying cluster identity of each feature.
}
\description{
This function is used to pool clustered features by adding the gene expression values within each cluster.
}
