% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{topic_cluster_QC_plot}
\alias{topic_cluster_QC_plot}
\title{QC plot for pooled data}
\usage{
topic_cluster_QC_plot(expr, clust.df, type = "genes")
}
\arguments{
\item{expr}{Expression data used for clustering.}

\item{clust.df}{Cluster matrix obtained with topic.cluster.matrix function.}

\item{type}{Set type to compare, options; "genes", "transcripts"}
}
\value{
Comparison plot between the number of unique genes in pooled clusters vs features.
}
\description{
This function is used to plot a comparison of the number of unique genes before and after pooling.
}
