% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{topic_clusters}
\alias{topic_clusters}
\title{Calculate clusters based on topics}
\usage{
topic_clusters(omega, method.dist = "euclidean", method.tree = "ward.D2",
  minClusterSize = 30)
}
\arguments{
\item{omega}{Topic data.frame.}

\item{method.dist}{Set distance method.}

\item{method.tree}{Set clustering method.}

\item{minCLusterSize}{Integer value specifying the minimum cluster size allowed.}
}
\value{
Integer vector specifying cluster identity of each feature.
}
\description{
This function is used to cluster features based on a topics matrix.
}
