% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{topic_heatmap}
\alias{topic_heatmap}
\title{Plot heatmap of lda results.}
\usage{
topic_heatmap(df, method.dist = "euclidean", method.tree = "ward.D2",
  minClusterSize = 30)
}
\arguments{
\item{df}{Topic data.frame (omega) or "topics" class object.}

\item{method.dist}{Set distance method.}

\item{method.tree}{Set clustering method.}

\item{minClusterSize}{Minimum cluster size.}
}
\value{
Heatmap of topic results.
}
\description{
This function is used to plot a heatmap of lda results.
}
