% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{var.genes}
\alias{var.genes}
\title{Detect variable genes using trendsceek package}
\usage{
var.genes(x, quantile.cutoff = 0.9, method = "glm", min.ncells.expr = 3,
  min.expr = 5)
}
\arguments{
\item{x}{spaceST object}

\item{quantile.cutoff}{set cutoff level}

\item{method}{method used for linear regression}
}
\description{
Find most variable genes using calc_varstats function from trendsceek.
}
