% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{var_genes}
\alias{var_genes}
\title{Detect variable genes using trendsceek package}
\usage{
var_genes(x, quantile.cutoff = 0.9, method = "glm",
  min.ncells.expr = 3, min.expr = 5)
}
\arguments{
\item{x}{spaceST object}

\item{quantile.cutoff}{set cutoff level}

\item{method}{method used for linear regression}

\item{min.ncells.expr}{An integer specifying the minimum number of cells a gene needs
to be expressed in with an expression level above 'min.expr'.}

\item{min.expr}{A numeric specifying the minimum expression required in a cell.}
}
\description{
Find most variable genes using calc_varstats function from trendsceek.
}
