% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_bootstrap.R
\name{violin.bootstraps}
\alias{violin.bootstraps}
\title{Violin plot of xCell scores}
\usage{
violin.bootstraps(df, type = "violin", cluster = NULL, group = NULL,
  samplename = NULL, ...)
}
\arguments{
\item{df}{Input data.frame.}

\item{type}{Charcter string specifying plot type ('violin' or 'boxplot').}
}
\value{
Plot of xCell score distributions.
}
\description{
Violin or boxplot of xCell scores.
}
