% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{volcano_plot_spaceST}
\alias{volcano_plot_spaceST}
\title{Volcano plot}
\usage{
volcano_plot_spaceST(object, cluster, FDR.cutoff = 0.05,
  logFC.cutoff = 1, annotate.top.genes = NULL, widget.out = NULL)
}
\arguments{
\item{object}{An object of class spaceST with differentially expressed genes.}

\item{cluster}{Select cluster to compare woith background.}

\item{FDR.cutoff}{Set cutoff threshold for adjusted p-values.}

\item{logFC.cutoff}{Set log2-foldchange cutoff.}

\item{annotate.top.genes}{Specify whether the top genes should be annotated in the plot.}

\item{widget.out}{Save html widget to file path.}
}
\description{
create a volcano plot of differentiually expressed genes
}
