% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_bootstrap.R
\name{xCell.bootstrap}
\alias{xCell.bootstrap}
\title{Bootstrap function for xCell}
\usage{
xCell.bootstrap(df, clusters, replicates, progressbar = T)
}
\arguments{
\item{df}{Input gene expression data.frame.}

\item{clusters}{Integer vector specifying cluster identity for each feature.}

\item{replicates}{Integer specifying number of replicates to run.}
}
\description{
Calculate xCell scores for bootstrap replicates of clusters.
}
