% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_tools.R
\name{xCell.heatmap}
\alias{xCell.heatmap}
\title{xCell heatmap}
\usage{
xCell.heatmap(df, group = NULL, scale = FALSE, compare = F,
  col.names = NULL, col2 = c("#4B0082", "#F0E442"))
}
\arguments{
\item{df}{data.frame with xCell output, i.e. cell types as rows and clusters as columns.}

\item{group}{Used to subset xCell data. Preset groups are; lymphoids, stem, stromal, myeloids, scores, others.}

\item{scale}{If TRUE, xCell scores will be scaled to a [0, 1] scale.}

\item{compare}{If TRUE, xCell scores will be scaled among all celltypes regardless of subtype group, making results
comparable between subtype groups.}

\item{col2}{Set "low" and "high" colors for heatmap.}
}
\value{
Heatmap of xCell scores.
}
\description{
Given an xCell output data.frame, this function can be used to generate a heatmap of xCell scores for celltypes
and clusters. You can subset the data using the default groups provided in the STanalysis3D package or provide a character vector
of cell type names. Clusters and cell types are automatically ordered with dendrograms using hclust (method = "Ward.D2").
}
