% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\name{batch.correct}
\alias{batch.correct}
\title{Method batch.correct for spaceST class.}
\usage{
batch.correct(object)
}
\arguments{
\item{object}{Object of class spaceST.}
}
\value{
Batch corrected data frame in slot "expr".
}
\description{
Function used to run a batch correction between samples.
}
\seealso{
\link[countClust]{BatchCorrectedCounts}
}
