% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST.R
\docType{methods}
\name{batch.correct}
\alias{batch.correct}
\alias{batch.correct,spaceST-method}
\title{Batch correction of spaceST data}
\usage{
batch.correct(object)

\S4method{batch.correct}{spaceST}(object)
}
\arguments{
\item{object}{object of class spaceST}
}
\value{
Batch corrected data frame in slot corrected
}
\description{
Run batch correction of replicate sections
}
