% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST.R
\name{batch.correct}
\alias{batch.correct}
\title{Batch correction of spaceST data}
\usage{
batch.correct(object)
}
\arguments{
\item{object}{spaceST object with expression data to correct}
}
\value{
Batch corrected data frame in slot corrected
}
\description{
Batch correction of spaceST data
}
