% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_coordinates}
\alias{get_coordinates}
\alias{get_coordinates.default}
\alias{get_coordinates.spaceST}
\alias{get_coordinates.dgCMatrix}
\title{get_coordinates: obtain spatial coordinates}
\usage{
get_coordinates(x, delimiter = "_")

\method{get_coordinates}{default}(x, delimiter = "_")

\method{get_coordinates}{spaceST}(x, delimiter = "_")

\method{get_coordinates}{dgCMatrix}(x, delimiter = "_")
}
\arguments{
\item{x}{List of expression data.frames, data.frame or matrix.}

\item{delimiter}{Specify character used to delimit coordinates in headers.}
}
\value{
Data frame with replicate numbers, x and y coordinates
}
\description{
This function is used to collect all coordinates from a list of expression datasets or an expression dataset
}
\details{
Obtain coordinates from expression data list of expression data.frame
}
