% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{CreatespaceSTobject}
\alias{CreatespaceSTobject}
\title{Filter ST expression data.}
\usage{
CreatespaceSTobject(raw.data, unique.genes = 500, min.exp = 2,
  min.features = 15, filter.data = RP, batch.correct = FALSE)
}
\arguments{
\item{raw.data}{A list of expression matrices/data.frames with genes as rows and feature coordinates as columns
Required structure of column names is: "replicate number_x_y"}

\item{unique.genes}{The lowest number of unique genes allowed in a feature}

\item{min.exp}{Integer value specifying the lowest expression value allowed at least min.features number of features (default 2)}

\item{min.features}{Integer value specifying the lowest number of features with at least min.exp (default 15)}

\item{filter.data}{A character vector specifying genes that should be filtered from the expression data (default RP)}

\item{batch.correct}{Logical specifying if data should be batch corrected}
}
\value{
A spaceST object containing filtered gene expression data for multiple replicates
}
\description{
This function is used to filter ST expression data used for 3D analysis. Features with low number of unique
genes and genes with low expression values will be removed. By default, ribosomal proteins and MALAT1 genes are also removed
}
