% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{DE_heatmap}
\alias{DE_heatmap}
\title{Plot DE heatmap}
\usage{
DE_heatmap(object, FDR.cutoff = 0.05, logFC.cutoff = 2, log.2 = T,
  type = "normalized", include.downregulated = F, scale.data = T)
}
\arguments{
\item{object}{spaceST object with DE results}

\item{FDR.cutoff}{cutoff for adjusted p-values}

\item{logFC.cutoff}{cutof for logFC}

\item{log.2}{log2 transform data}

\item{type}{select data type from spaceST object}

\item{scale.data}{logical speficying whether or not values should be scaled}
}
\value{
Heatmap of DE genes.
}
\description{
Function used to plot heatmaps of DE genes expression
}
