% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST.R
\name{normalize}
\alias{normalize}
\title{Normalize ST data}
\usage{
normalize(object, method = "cp10k", log2 = F)
}
\arguments{
\item{object}{spaceST object with expression data to correct}

\item{method}{select normalization method [default "cp10k"]. Curently only sp10k available}

\item{log2}{Logical specifying whether or not data should be log2-transformed}
}
\value{
Normalized data frame in slot normalized
}
\description{
Normalize ST data
}
