% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\name{NormalizespaceST}
\alias{NormalizespaceST}
\title{Method NormalizespaceST ST data.}
\usage{
NormalizespaceST(object, method = "scran", log2 = F, clusters = NULL, ...)
}
\arguments{
\item{object}{Object of class spaceST.}

\item{method}{Select normalization method, default "cp10k" [options: "cp10k", "scran"].}

\item{log2}{Logical specifying whether or not data should be log2-transformed.}

\item{clusters}{Integer vector specifying clusters used for normalization with scran on large gene expression matrices.
By default, a clustering step is performed using \link[scran]{quickCluster} if there are more than 500 array spots in the
expression matrix.}
}
\value{
Normalized expression data.
}
\description{
Normalization of gene expression data using either counts per 10k or scran.
}
\seealso{
\link[scater]{Normalize} and \link[scran]{computeSumFactors}
}
