% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tesselation.R
\name{palette.select}
\alias{palette.select}
\title{Create palettes}
\usage{
palette.select(palette)
}
\arguments{
\item{palette}{select palette [options: "GnBu", "RdBu", "the.cols", "Spectral", "offwhite.to.black", "viridis", "cividis", "magma", "plasma", "heat"]}
}
\description{
returns a palette using the coloRamp function
}
