% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{RunTSNEspaceST}
\alias{RunTSNEspaceST}
\title{t-SNE of spaceST data}
\usage{
RunTSNEspaceST(object, dims = 2, initial.dims = 50, theta = 0,
  check_duplicates = FALSE, pca = TRUE, perplexity = 15,
  max_iter = 1000, verbose = FALSE, seed = 0, plot.tSNE = FALSE,
  use.dims = "topics", select.dims = NULL, clusters = NULL)
}
\arguments{
\item{object}{A spaceST object}

\item{seed}{Set seed for reproducibility}

\item{plot.tSNE}{Logical specifying whether or not t-SNE results should be plotted}

\item{use.dims}{Specify what dimensions should be used for t-SNE (currently only "topics" available)}

\item{clusters}{Integer vector used to color features in t-SNE plot}
}
\value{
Matrix with t-SNE results
}
\description{
Run t-SNE on spaceST object. Results are saved in the tsne slot.
}
