% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{scatter_HE}
\alias{scatter_HE}
\title{Generate binary dot pattern}
\usage{
scatter_HE(img, coords = NULL, limit = 0.5, rownum = 50000,
  show.plot = FALSE, offset_x = 0.75, offset_y = 0.75)
}
\arguments{
\item{img}{File path or list of file paths for BW image(s).}

\item{limit}{Defines cutoff on the greyscale for what to interpret as a valid coordinate.}

\item{rownum}{Number of rows in output grid, i.e. the maximum number of points to keep.}
}
\value{
Data.frame with x, y coordinates representing cell nuclei on an HE image.
}
\description{
Given a black and white HE image, this function can be used to create a scatter representatio of cell nuclei ditributed
over the tissue image. The image needs to be completely free from noise surrounding the tissue area and make sure that the colorspace
is set to grey.
}
\examples{
library(spaceST)
# Create scatter pattern and plot results

scatter <- scatter_HE(img = <path to black and white HE image>, show.plot = T)
}
