% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST.R
\docType{methods}
\name{show}
\alias{show}
\alias{spaceST-show}
\title{show method for spaceST}
\usage{
\S4method{show}{spaceST}(object)
}
\arguments{
\item{object}{A spaceST object}
}
\description{
show method for spaceST
}
