% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\docType{methods}
\name{spaceST-methods}
\alias{spaceST-methods}
\alias{batch.correct}
\alias{batch.correct,spaceST-methods}
\alias{NormalizespaceST}
\alias{NormalizespaceST,spaceST-methods}
\alias{plot.QC.spaceST}
\alias{plot.QC.spaceST,spaceST-methods}
\alias{spPCA}
\alias{spPCA,spaceST-methods}
\alias{spots.under.tissue}
\alias{spots.under.tissue,spaceST-methods}
\alias{filter}
\alias{filter,spaceST-methods}
\alias{dim}
\alias{dim,spaceST-methods}
\alias{counts}
\alias{counts,spaceST-methods}
\alias{normcounts}
\alias{normcounts,spaceST-methods}
\alias{topic.clusters}
\alias{topic.clusters,spaceST-methods}
\title{space-ST methods.}
\usage{
\S4method{batch.correct}{spaceST}(object)

\S4method{NormalizespaceST}{spaceST}(object, method = "cp10k", log2 = F,
  pcount = 1, cluster = NULL, ...)

\S4method{plot.QC.spaceST}{spaceST}(object, separate = F)

\S4method{spPCA}{spaceST}(object, ntop = 500, ncomponents = 2,
  exprs_values = "norm.data", ...)

\S4method{spots.under.tissue}{spaceST}(object, selection.files,
  delimiter = "_")

\S4method{filter}{spaceST}(object, unique.genes = 300, min.exp = 2,
  min.features = 15, filter.genes = NULL, force.filter = FALSE,
  delimiter = "_")

\S4method{dim}{spaceST}(x)

\S4method{counts}{spaceST}(object)

\S4method{normcounts}{spaceST}(object)

\S4method{topic.clusters}{spaceST}(object)
}
\arguments{
\item{object}{Object of class spaceST.}
}
\description{
space-ST methods.

Method dim
}
