% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST.R
\docType{class}
\name{spaceST}
\alias{spaceST}
\alias{spaceST-class}
\title{The spaceST class}
\description{
An S4 class object used to filter and store spatial transcriptomics data from
multiple tissue sections
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{Raw input data}

\item{\code{expr}}{A data.frame with filtered gene expression data}

\item{\code{corrected}}{A data.frame with batch corrected gene expressiond data}

\item{\code{normalized}}{A matrix with normalized expression values}

\item{\code{topics}}{Topic matrix obtained using compute.lda() function from cellTree}

\item{\code{coordinates}}{A data.frame to represent feature coordinates in the expression dataset}

\item{\code{filter.settings}}{A list with settings used for filtering}

\item{\code{statistics}}{A list with a data.frame and plot object for quality control}
}}

