% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_tools.R
\name{spatial.clusters}
\alias{spatial.clusters}
\title{Visualize spatial distribution of clusters}
\usage{
spatial.clusters(clusters, coords, HE, size = 4, xlim = c(0, 32),
  ylim = c(0, 34), type = "ST_grid", ...)
}
\arguments{
\item{clusters}{Integer vector specifying feature cluster identity for corresponding feature coordinates.}

\item{coords}{data.frame with x, y values for feature coordinates.}

\item{HE}{Path to cropped black and white HE image.}

\item{size}{numerical specifying coordinate size.}

\item{...}{additional parameters passed to geom_point(), defining the attributes of the feature coordinate points.}
}
\value{
Plot of clustered featured overlayed on black and white HE image.
}
\description{
Given a black and white HE image, aligned feature coordinates and a cluster vector, this function
can be used to visualize the spatial distribution of clusters.

NOTE: The input image should be cropped to the centers of the frame corners, i.e. the outermost layer of features.
This is a requirement for the
}
\examples{
library(spaceST)

data(bcST)
spaceST.object <- CreatespaceSTobject(bcST, corrected = T)

# Run factor analysis and cluster features
spaceST.object <- topic_compute(spaceST.object)
clusters <- topic_clusters(topic.df)

# Get coordinates from spaceST object
coords <- get_coordinates(spaceST)[, 1:2]

# Plot spatial heatmap
spatial.clusters(clusters, coords, HE = "/path/to/HE")
}
