% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_tools.R
\name{spatial.clusters}
\alias{spatial.clusters}
\title{Visualize spatial distribution of clusters.}
\usage{
spatial.clusters(object, clusters = NULL, HE.list = NULL,
  arrange = T, ncols = NULL, xlim = c(1, 33), ylim = c(1, 35),
  cols = NULL, ...)
}
\arguments{
\item{object}{Object of class spaceST.}

\item{clusters}{Integer/numeric vector specifying clusters.}

\item{HE.list}{List of paths to HE images in jpeg format that should be used as a background for the
spatial heatmap.}

\item{arrange}{Arrange plots.}

\item{ncols}{Number of columns in arranged plot table.}

\item{xlim, ylim}{Set limits of x/y axes. [default: xlim = c(1, 34), ylim = c(1, 36)]}

\item{cols}{Set custom cluster colors by specifying a character vector with color codes. You can specify
specific colors for each cluster by naming the character vector with the cluster ids.}

\item{...}{additional parameters passed to geom_point(), defining the attributes of the feature coordinate points.}
}
\value{
Plot of clustered featured overlayed on black and white HE image.
}
\description{
Visualize spatial distribution of clusters.
}
