% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceST_methods.R
\name{spPCA}
\alias{spPCA}
\title{Method spPCA for spaceST class.}
\usage{
spPCA(object, ntop = 500, exprs_values = "norm.data", ...)
}
\arguments{
\item{object}{Object of class spaceST.}

\item{ntop}{Number of variable genes to select for PC analysis.}

\item{exprs_values}{Select target object to use as input for PC analysis [options: "norm.data", "expr"]}

\item{...}{Parameters passed to \link[stats]{prcomp}.}
}
\description{
Run PC analysis on spaceST object.
}
\seealso{
\link[stats]{prcomp}
}
