% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tesselation.R
\name{tesselation.plot}
\alias{tesselation.plot}
\title{Voronoi tesselation plot}
\usage{
tesselation.plot(r, datapolyM = NULL, outline.color = NA, bg = "black",
  xlim = c(0, 34), ylim = c(0, 36), alpha = 1)
}
\arguments{
\item{r}{data.frame with grouped x, y values used with geom_polygon}

\item{datapolyM}{convex hull polygon used to mask area outside tissue}

\item{outline.color}{set color of polygon outline}

\item{bg}{background color}

\item{xlim}{set limits of x axis}

\item{ylim}{set limits of y axis}
}
\description{
Compute voronoi tesselation and visualize ST data as a colored mosaic
}
