% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{topic_compute}
\alias{topic_compute}
\title{Helper function to compute lda using cellTree package
TODO: fix bbug in Set dataset "raw" for single df}
\usage{
topic_compute(object, min.topic = 2, max.topic = 15, num.genes = NULL,
  type = "corrected", method = "maptpx", sd.filter = F, log.scale = F,
  force.recalc = F, minClusterSize = 30)
}
\arguments{
\item{object}{spaceST object with expression data or data.frame}

\item{max.topic}{Maximum number of topics}

\item{num.genes}{Number of genes to use for lda modelling}

\item{type}{Character string specifying if the "normalized", "corrected" or "raw" data should be used as input. Default
is set to "corrected"}

\item{method}{method passed to CellTree function}

\item{sd.filter}{Standard deviation threshold}

\item{log.scale}{Convert to log scale}

\item{min.topics}{Minimum number of topics}
}
\value{
Topic matrix
}
\description{
This helper function is used to compute topics for an expression dataset using lda modelling
from cellTree package
}
