% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{topic_compute}
\alias{topic_compute}
\alias{topic_compute.default}
\alias{topic_compute.spaceST}
\title{Helper function to compute lda using cellTree package}
\usage{
topic_compute(object, min.topic = 2, max.topic = 15, num.genes = NULL,
  type = "corrected", method = "maptpx", sd.filter = F, log.scale = F)

\method{topic_compute}{default}(object, min.topic = 2, max.topic = 15,
  num.genes = NULL, method = "maptpx", sd.filter = F, log.scale = F)

\method{topic_compute}{spaceST}(object, min.topic = 2, max.topic = 15,
  num.genes = NULL, method = "maptpx", sd.filter = FALSE,
  log.scale = FALSE, force.recalc = FALSE)
}
\arguments{
\item{object}{spaceST object with expression data or data.frame}

\item{max.topic}{Maximum number of topics}

\item{num.genes}{Number of genes to use for lda modelling}

\item{type}{Character string specifying if the "normalized", "corrected" or "raw" data should be used as input. Default
is set to "corrected"}
}
\value{
Topic matrix
}
\description{
This helper function is used to compute topics for an expression dataset using lda modelling
from cellTree package
}
