% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_functions.R
\name{topic_heatmap}
\alias{topic_heatmap}
\title{Plot heatmap of lda results.}
\usage{
topic_heatmap(object, method.dist = "euclidean", method.tree = "ward.D2",
  minClusterSize = 30, cols = NULL)
}
\arguments{
\item{object}{Object of class spaceST.}

\item{method.dist}{Set distance method.}

\item{method.tree}{Set clustering method.}

\item{minClusterSize}{Minimum cluster size.}

\item{cols}{Set cluster colors.}
}
\value{
Heatmap of topic results.
}
\description{
This function is used to plot a heatmap of lda results.
}
