% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{volcano_plot_spaceST}
\alias{volcano_plot_spaceST}
\title{Volcano plot}
\usage{
volcano_plot_spaceST(object, cluster, FDR.cutoff = 0.05, logFC.cutoff = 1,
  annotate.top.genes = NULL, widget.out = NULL)
}
\arguments{
\item{object}{spaceST object with differentially expressed genes}

\item{FDR.cutoff}{set cutoff threshold for adjusted p-values}

\item{logFC.cutoff}{set log2-foldchange cutoff}

\item{annotate.top.genes}{specify whether the top genes should be annotated in the plot}

\item{widget}{save html widget to file path}
}
\description{
create a volcano plot of differentiually expressed genes
}
