% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell_bootstrap.R
\docType{class}
\name{xCellBoot}
\alias{xCellBoot}
\alias{xCellBoot-class}
\title{xCellBoot class}
\description{
xCellBoot class
}
\section{Slots}{

\describe{
\item{\code{rep.list}}{A list of pseudo-replicate bootstrap results}

\item{\code{obs.df}}{Data.frame with observed xCell values}

\item{\code{mean.df}}{Data.frame with median bootstrap values}

\item{\code{bias.df}}{Data.frame with bias values, i.e. difference between observed and median values}

\item{\code{betadist}}{p-values}
}}

